refactor for import
