+++ back-end/api/wsgi.py
@@ -1,6 +1,6 @@
 ### This is the entry point for the application, run by gunicorn server.
 
-from myproject import app
+from apiapp import app
 
 if __name__ == "__main__":
     app.run()
